+++ torch/onnx/symbolic_helper.py
@@ -173,12 +173,12 @@ def parse_args(*arg_descriptors):
         @wraps(fn)
         def wrapper(g, *args, **kwargs):
             # some args may be optional, so the length may be smaller
-            FILE_BUG_MSG = "If you believe this is not due to custom symbolic implementation within your code or"\
+            FILE_BUG_MSG = "If you believe this is not due to custom symbolic implementation within your code or "\
                 "an external library, please file an issue at "\
-                "https://github.com/pytorch/pytorch/issues/new?template=bug-report.md to report this bug."
+                "https://github.com/pytorch/pytorch/issues/new?template=bug-report.yml to report this bug."
             assert len(arg_descriptors) >= len(args),\
-                f"A mismatch between the number of arguments ({len(args)}) and"\
-                f" their descriptors ({len(arg_descriptors)}) was found at symbolic function '{fn.__name__}'. "\
+                f"A mismatch between the number of arguments ({len(args)}) and "\
+                f"their descriptors ({len(arg_descriptors)}) was found at symbolic function '{fn.__name__}'. "\
                 f"{FILE_BUG_MSG}"
 
             try:
